Moving remaining tests out of integration
