+++ Audit.NET.sln
@@ -160,6 +160,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.Serilog.UnitTest", "t
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.AmazonQLDB.UnitTest", "test\Audit.AmazonQLDB.UnitTest\Audit.AmazonQLDB.UnitTest.csproj", "{ED5B6C4F-87D2-46AA-BA41-3BA4B0B1F551}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Audit.Wcf.Client.UnitTest", "test\Audit.Wcf.Client.UnitTest\Audit.Wcf.Client.UnitTest.csproj", "{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -696,6 +698,14 @@ Global
 		{ED5B6C4F-87D2-46AA-BA41-3BA4B0B1F551}.Release|Any CPU.Build.0 = Release|Any CPU
 		{ED5B6C4F-87D2-46AA-BA41-3BA4B0B1F551}.Release|x64.ActiveCfg = Release|Any CPU
 		{ED5B6C4F-87D2-46AA-BA41-3BA4B0B1F551}.Release|x64.Build.0 = Release|Any CPU
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}.Debug|x64.Build.0 = Debug|Any CPU
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}.Release|x64.ActiveCfg = Release|Any CPU
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -769,6 +779,7 @@ Global
 		{E3821A4F-B1F5-4220-A491-4310D4DC4DE4} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 		{25E2A9F4-DB37-437D-8688-DDB5C678C0F8} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 		{ED5B6C4F-87D2-46AA-BA41-3BA4B0B1F551} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
+		{B0EDF983-3AC1-4D8D-9897-A7B639A0001F} = {A54B4BB6-3439-432B-AFD9-FE62D6528D42}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B1989475-43D6-4AA5-9717-6DBF734B0C6E}
