+++ test/Audit.Wcf.Client.UnitTest/CatalogService.cs
