+++ test/Audit.Wcf.Client.UnitTest/ICatalogService.cs
