+++ test/Audit.Wcf.Client.UnitTest/WcfClientTests.cs
