+++ test/Run-Tests.ps1
@@ -33,7 +33,8 @@ StartDotnetUnitTests 'Audit.JsonAdapter.UnitTest' 'JsonAdapter';
 StartDotnetUnitTests 'Audit.WebApi.UnitTest' 'WebApi';
 StartDotnetUnitTests 'Audit.DynamicProxy.UnitTest' 'DynamicProxy';
 StartDotnetUnitTests 'Audit.Redis.UnitTest' 'Redis';
-StartDotnetUnitTests 'Audit.Wcf.UnitTest' 'Wcf';
+StartDotnetUnitTests 'Audit.Wcf.UnitTest' 'Wcf.Server';
+StartDotnetUnitTests 'Audit.Wcf.Client.UnitTest' 'Wcf.Client';
 StartDotnetUnitTests 'Audit.RavenDB.UnitTest' 'RavenDB';
 StartDotnetUnitTests 'Audit.SqlServer.UnitTest' 'SqlServer';
 StartDotnetUnitTests 'Audit.IntegrationTest' 'Integration' '--filter=TestCategory!=AzureDocDb&TestCategory!=AzureBlob&TestCategory!=AzureStorageBlobs&TestCategory!=WCF&TestCategory!=Elasticsearch&TestCategory!=Dynamo&TestCategory!=PostgreSQL&TestCategory!=Kafka&TestCategory!=AmazonQLDB&TestCategory!=Mongo&TestCategory!=MySql';
